% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smo.R
\name{fit_smo}
\alias{fit_smo}
\title{SMO outerloop function to select i2}
\usage{
fit_smo(model, max_iterations = 10000)
}
\arguments{
\item{model}{SmoModel}

\item{max_iterations}{(default 10000)}
}
\value{
SmoModel
}
\description{
SMO outerloop function to select i2
}
\examples{

toy_x <- iris |>
  dplyr::select(-c(Species))
toy_y <- iris |>
  dplyr::mutate(y = dplyr::if_else(Species == "setosa", 1, -1)) |>
  dplyr::pull(y)
toy_model <- SmoModel$new(x = toy_x, y = toy_y, kernel = "linear")
fit_smo(toy_model)

}
