% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smo.R
\name{SmoModel}
\alias{SmoModel}
\title{Sequential Minimal Optimization model}
\description{
R6Class to maintain references to SMO elements.
Includes functions to initialize, print, and finalize (delete).
}
\examples{

toy_x <- iris |>
  dplyr::select(-c(Species))
toy_y <- iris |>
  dplyr::mutate(y = dplyr::if_else(Species == "setosa", 1, -1)) |>
  dplyr::pull(y)
toy_model <- SmoModel$new(x = toy_x, y = toy_y, kernel = "linear")

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{x}}{matrix or data.frame of observations}

\item{\code{y}}{data.frame, matrix, or vector of target values}

\item{\code{kernel}}{name}

\item{\code{b}}{bias}

\item{\code{beta}}{for sigmoid or gaussian kernel}

\item{\code{theta}}{for sigmoid kernel}

\item{\code{d}}{for polynomial kernel}

\item{\code{epsilon}}{sensitivity of margin to errors}

\item{\code{tol}}{convergence tolerance parameter}

\item{\code{C}}{margin}

\item{\code{alpha}}{vector}

\item{\code{E}}{error cache}

\item{\code{verbose}}{TRUE to print messages}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SmoModel-new}{\code{SmoModel$new()}}
\item \href{#method-SmoModel-print}{\code{SmoModel$print()}}
\item \href{#method-SmoModel-finalize}{\code{SmoModel$finalize()}}
\item \href{#method-SmoModel-K}{\code{SmoModel$K()}}
\item \href{#method-SmoModel-dual}{\code{SmoModel$dual()}}
\item \href{#method-SmoModel-f}{\code{SmoModel$f()}}
\item \href{#method-SmoModel-calculateE}{\code{SmoModel$calculateE()}}
\item \href{#method-SmoModel-takeStep}{\code{SmoModel$takeStep()}}
\item \href{#method-SmoModel-boundAlphaIndex}{\code{SmoModel$boundAlphaIndex()}}
\item \href{#method-SmoModel-examineExamples}{\code{SmoModel$examineExamples()}}
\item \href{#method-SmoModel-predict}{\code{SmoModel$predict()}}
\item \href{#method-SmoModel-clone}{\code{SmoModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmoModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-SmoModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize an SmoModel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmoModel$new(
  x,
  y,
  kernel = "linear",
  b = 0,
  beta = 1,
  d = 1,
  sigma = 1,
  theta = 1,
  epsilon = 10^(-3),
  tol = 10^(-3),
  C = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{matrix or data.frame of observations}

\item{\code{y}}{data.frame, matrix, or vector of target values}

\item{\code{kernel}}{name}

\item{\code{b}}{bias}

\item{\code{beta}}{for sigmoid or gaussian kernel}

\item{\code{d}}{for polynomial kernel}

\item{\code{sigma}}{for gaussian kernel}

\item{\code{theta}}{for sigmoid kernel}

\item{\code{epsilon}}{sensitivity of margin to errors}

\item{\code{tol}}{convergence tolerance parameter}

\item{\code{C}}{margin}

\item{\code{verbose}}{TRUE to print messages}

\item{\code{alpha}}{vector}

\item{\code{E}}{error cache}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmoModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-SmoModel-print}{}}}
\subsection{Method \code{print()}}{
print information about the instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmoModel$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmoModel-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-SmoModel-finalize}{}}}
\subsection{Method \code{finalize()}}{
call to delete instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmoModel$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmoModel-K"></a>}}
\if{latex}{\out{\hypertarget{method-SmoModel-K}{}}}
\subsection{Method \code{K()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmoModel$K(x, y = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmoModel-dual"></a>}}
\if{latex}{\out{\hypertarget{method-SmoModel-dual}{}}}
\subsection{Method \code{dual()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmoModel$dual(alpha)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmoModel-f"></a>}}
\if{latex}{\out{\hypertarget{method-SmoModel-f}{}}}
\subsection{Method \code{f()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmoModel$f(i, v = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmoModel-calculateE"></a>}}
\if{latex}{\out{\hypertarget{method-SmoModel-calculateE}{}}}
\subsection{Method \code{calculateE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmoModel$calculateE(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmoModel-takeStep"></a>}}
\if{latex}{\out{\hypertarget{method-SmoModel-takeStep}{}}}
\subsection{Method \code{takeStep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmoModel$takeStep(i1, i2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmoModel-boundAlphaIndex"></a>}}
\if{latex}{\out{\hypertarget{method-SmoModel-boundAlphaIndex}{}}}
\subsection{Method \code{boundAlphaIndex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmoModel$boundAlphaIndex(bound, index_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmoModel-examineExamples"></a>}}
\if{latex}{\out{\hypertarget{method-SmoModel-examineExamples}{}}}
\subsection{Method \code{examineExamples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmoModel$examineExamples(i2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmoModel-predict"></a>}}
\if{latex}{\out{\hypertarget{method-SmoModel-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmoModel$predict(v)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmoModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SmoModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmoModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
